# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Yalla Wassel delivery management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zones:
    get:
      operationId: listZones
      tags: [zones]
      summary: List all active zones
      parameters:
        - name: governorate
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"
  /zones/{id}:
    get:
      operationId: getZone
      tags: [zones]
      summary: Get zone by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
        "404":
          description: Not found

  # ── Users / Drivers ──────────────────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (manager only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [manager, driver, customer]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/drivers:
    get:
      operationId: listDrivers
      tags: [users]
      summary: List all drivers with their stats
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [available, busy, off]
        - name: zoneId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DriverWithStats"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [users]
      summary: Update driver status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Orders ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Zone:
      type: object
      required: [id, nameAr, nameEn, governorate, region, isActive]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        governorate:
          type: string
        region:
          type: string
        isActive:
          type: boolean

    # ── User schemas ───────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, fullName, phone, role, status, isActive, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [manager, driver, customer]
        zoneId:
          type: ["integer", "null"]
        zoneName:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, busy, off]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    DriverWithStats:
      type: object
      required: [id, fullName, phone, status, isActive, createdAt, totalOrders, deliveredToday, avgRating]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        zoneId:
          type: ["integer", "null"]
        zoneName:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, busy, off]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        totalOrders:
          type: integer
        deliveredToday:
          type: integer
        avgRating:
          type: number
        isTrusted:
          type: boolean
    UserStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [available, busy, off]

    # ── Order schemas ──────────────────────────────────────────────────────────
...